Refact: 사용자의 숫자야구 응답 검증 코드 클래스로 분리
